refactor(arch): rename bytecode/arch.py to bytecode/grid.py with shim (stage 4 of #569)
Fourth stage of the multi-stage package merge tracked in #569. The
old bytecode/arch.py held only two Grid conversion utilities
(grid_to_rust, grid_from_rust) — the filename didn't match its
responsibility, and the slot is needed for the higher-tier arch
surface in later stages.

Renamed to bytecode/grid.py. Old path bytecode/arch.py remains as a
re-export shim so anything that imported grid_to_rust/grid_from_rust
from the old location keeps working. The shim is removed in the
final cleanup stage.

Refs #569.
