+++ python/bloqade/lanes/bytecode/arch.py
@@ -1,23 +1,13 @@
-from bloqade.geometry.dialects import grid as geometry_grid
+"""TRANSITIONAL SHIM — see ``.superpowers/plans/2026-04-27-archspec-package-merge.md``
+(Stage 4) for the rationale. The original ``bytecode/arch.py`` only held
+two ``Grid`` conversion utilities; the file was renamed to ``grid.py``
+(matching its responsibility) to free the ``arch`` slot for the higher-tier
+arch surface.
 
-from bloqade.lanes.bytecode._native import Grid
+Removed in the final cleanup stage.
+"""
 
-
-def grid_to_rust(g: geometry_grid.Grid) -> Grid:
-    """Convert a bloqade-geometry Grid to a Rust bytecode Grid."""
-    return Grid(
-        x_start=g.x_init,  # type: ignore[arg-type]
-        y_start=g.y_init,  # type: ignore[arg-type]
-        x_spacing=list(g.x_spacing),
-        y_spacing=list(g.y_spacing),
-    )
-
-
-def grid_from_rust(g: Grid) -> geometry_grid.Grid:
-    """Convert a Rust bytecode Grid to a bloqade-geometry Grid."""
-    return geometry_grid.Grid(
-        x_init=g.x_start,
-        y_init=g.y_start,
-        x_spacing=list(g.x_spacing),  # type: ignore[arg-type]
-        y_spacing=list(g.y_spacing),  # type: ignore[arg-type]
-    )
+from bloqade.lanes.bytecode.grid import (
+    grid_from_rust as grid_from_rust,
+    grid_to_rust as grid_to_rust,
+)
